Correct the write completion status expection in zap_test
When server writes to a map that has been unmapped by client, the status
of the write completion event could be either ZAP_ERR_REMOTE_MAP or
ZAP_ERR_OK depending on the underlying transport. The patch fixes the
zap_test program so that it verifies the write completion status to the
correct ZAP error value.