Introduce ldms_set_light_copy() to the ldms library
ldms_set_light_copy() makes a snapshot of an LDMS set. The content of
the snapshot set is constant, regardless of any changes to the original
set. If the array cardinality of the source set is 1, the snapshot set
is exactly the same as the original set at the time of the snapshot. If
the array cardinality is greater than one, the snapshort set has only
the metric data of the current index in the array.