Modify how ovis_log_set_level_by_name() handle a non-existing logger
Without the patch, ovis_log_set_level_by_name() returns ENOENT if there is no
logger with the given name. This prevents applications from setting the log
level of a logger before it has been registered. Specifically for LDMSD, admins
cannot set the log level of Zap-related logger in configuration files. LDMSD
handles the log_level command before it listens on a port, and the Zap
libraries are loaded at the listen time.

A caveat is that only the logger names can set the log level if the loggers do
not exist yet. Regular expressions do not work because there are no names to
match.