Change the producer set' state to DELETING in the deleting path
The patch is an effort to prevent LDMSD posting a lookup request or an update request while
the producer set is being deleted. It addresses three things.

 * It renames the producer state DELETED to DELETING to better represent
the producer set.
 * It makes the prdcr_set_reset() function change the producer set's
state to DELETING. It is to prevent the daemon from sending an update
request while it is deleting the set.
 * It makes the prdcr_set_del() function change the producer set's state
to DELETING rather than START. This is to prevent the daemon to send a
lookup request to the server while it is deleting the set.  LDMSD
submits a lookup request if the producer set is in the START state.