Use  peer's hostname and listen port for advertised producer's names
Currently, an advertised producer's name is the name of the advertiser
on the peer. This creates confusion. The patch constructs the producer
name as <hostname>:<port>, where hostname is the peer hostname and port
is the listening port. If there are multiple listening ports, ldmsd uses
the first listening port.