Fix __APPEND macro to prevent infinite loop of reallocation
The macro was causing an indefinite loop when the length of the string
to be appended equaled the size of the remaining buffer. In this case,
LDMS_ROUNDUP(len-sz, __APPEND_SZ) evaluted to zero, resulting in
reallocation with the same buffer size repeatedly.

This patch ensures that the buffer is always reallocated with a larger
size when len == sz, preventing the infinite loop condition.