Add ldms_ls_to_test_sampler.py script
ldms_ls_to_test_sampler.py is a Python script that takes ldms_ls -vl
output and convert it to test_sampler config lines.

The script needs to be updated to include list and record as well as to
be compatible with the current ldms_ls -vl output format.

The ldms_ls.out file is included in the patch as a reference of the
ldms_ls -vl output that the script is compatiblie with.