Fix rst2man output
The output of rst2man in `rules.mk` were modified to address indentation
issues probably from the older version of rst2man. However, these
modifications incorrectly formatted the the recent rst2man output
(`python3-docutils` on Ubuntu 22.04). For example, with the modification
the paragraph indentation was not resetted after a list has ended.

INPUT:
```
This is a list.

- one
- two
- three

And so on.

```

We should expect the following output:

```
This is a list.

* one
* two
* three

And so on.
```

But, instead we got (with the mod):
```
This is a list.

* one
* two
* three

  And so on.
```

Removing the modifications solved the formatting issue.